fixed issues regarding success rates

+++ Source/VCFarming/Public/FarmLand.h
@@ -20,7 +20,7 @@ public:
 	UFUNCTION(BlueprintCallable)
 	void InputCropType(int index, int SuccessRate);
 
-	int CurrentSuccessRate;
+	float CurrentSuccessRate;
 protected:
 	// Called when the game starts or when spawned
 	virtual void BeginPlay() override;
