+++ Source/VCFarming/Public/FarmLand4.h
@@ -33,7 +33,7 @@ public:
 	
 	void InputCropType(int index, int SuccessRate);
 
-	int CurrentSuccessRate;
+	float CurrentSuccessRate;
