+++ Source/VCFarming/Public/FarmLand5.h
@@ -31,7 +31,7 @@ public:
 	
 	void InputCropType(int index, int SuccessRate);
 
-	int CurrentSuccessRate;
+	float CurrentSuccessRate;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	TArray<UStaticMesh*> CropMeshes;
